Refactor ClientRepository and add randomGenerator.h
Added randomGenerator.h to project files and clientRepositoryTest.
Refactored ClientRepository to accept individual client details.
Overloaded add and update methods for flexibility.
Changed add method return type to size_t in IRepository.h.
Updated clientRepositoryTest to use new methods.

+++ Domain/Repositories/IRepository.h
@@ -5,7 +5,7 @@ template<typename TReturn>
 class IRepository
 {
 public:
-	virtual int add(const TReturn* class_) = 0;
+	virtual size_t add(const TReturn* class_) = 0;
 	virtual TReturn* get(size_t id) = 0;
 	virtual bool update(const TReturn* class_) = 0;
 	virtual bool deleteClass(size_t id) = 0;
